backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
motor>=3.3.2
pymongo>=4.6.0
dnspython>=2.6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
numpy>=1.26.0
scikit-learn>=1.3.2
pandas>=2.0.3
joblib>=1.3.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
certifi>=2024.0.0
